Missing QTR length samples are filled using previous years QTR

Fills in missing length samples for YEAR/QTR combinations by using the previous YEARs value for the same QTR. If this is also missing, then the previous YEARs value for the same QTR is used and so on. For YEARs where there are landings before any length samples were taken all YEAR/QTRs are assigned the length samples from the most recent YEAR where length samples were taken (from the same QTR)

aggregate_to_qtr(
  data,
  gearType,
  marketCode,
  QTRData,
  missingEarlyYears,
  nLengthSamples,
  pValue,
  outputDir,
  logfile
)

Arguments

data
List. Landings data and length data

gearType
Character string. NEGEAR gear code

marketCode
Character string. MARKET_CODE designation from cfdbs

QTRData
Tibble. (n x 8). Filtered landings data by NEGEAR, MARKET_CODE, YEARs > earlyYears

missingEarlyYears
Numeric vector. Years prior to first length sample was taken

pValue
Numeric scalar. Threshold pvalue for determining significance of ks test for length samples

outputDir
Character string. Path to output directory (png files saved here)

nLenthSampels
Numeric scalar. Number of length samples deemed to be "enough" for calculations. This is passed from aggregate_landings

logFile
Character string. Specify the name for the log file generated describing all decisions made.

Value

List. Same as input data