Fit length weight relationship to species data from SVDBS

Fits a length weight relationship for use in length expansion

fit_length_weight(lengthWeightData, speciesName, outputDir, logfile)

Arguments

lengthWeightData
Data frame. length-weight pairs. Each row represents an individual fish

speciesName
Character string. Common name for species

Value

List of model fit objects

commonSlope

lm object. Fit for single slope (beta)

seasonalSlope

lm object. Fit for seasonal slopes

Notes on model fitting

The simplest Null model (H0) is assumed to be $$W_i = \alpha L_i^\beta exp(e_i)$$

where W_i = Weight and L_i = Length of fish i, e_i ~ \(N(0,\sigma^2)\) and \(\alpha\) & \(\beta\) are intercept and slope parameters (on the log scale) to be estimated.

The alternative H1: \(\beta_j != \beta\) where \(\beta_j\) is the slope parameter (on the log scale) for season i. i = 1, ..., 4 (spring, summer, fall, winter)

The above hypothesis is tested and the pvalue is output in the log file. Plots of model fits are also produced and saved in the output directory