Get species object

Lists set of rules required for catch at length aggregation

get_species_object(species_itis = NULL, stock = NA)

Arguments

species_itis

Numeric. Itis code for species (Default = NA, Returns a NA object for user to define)

stock

Character string. Define stock region for species_itis (Default = NA)

Value

A List

species_itis

6 digit species ITIS code

speciesName

Species Name

SVSPP

Species survey SVSPP code

temporalAggregation

Level of temporal aggregation: "QTR","SEMESTER","YEAR"

marketCodes

Data frame: mapping Market codes to use with market codes to combine

gearCodes

Data frame: mapping gear codes to use with gear codes to combine

statStockArea

Character array of statistical areas (3 characters)

howAggregate

How to aggregate the landings data in absense of length samples: "borrow","combine"

LengthWeightData

Where length-weight data comes from: "survey","commercial","both"

LengthWeightRelationships

How to fit length-weight relationships. By: "semester","quarter"

LengthWeightTimeBlocks

Time block over which to use for length weight data: "all", c(years)

AgeData

Where Age data from from: "survey","commercial"

AgeLengthKey

Variables over which to calculate age length key: "semester","qtr","gear"

startDate

First year of assessment

stock

Defines the stock, if more than one stock for a species:"GB","GOM","North","South